secured routes with admin check

+++ server/api/products.js
@@ -2,7 +2,27 @@ const router = require("express").Router();
 const {
   models: { Product },
 } = require("../db");
+const jwt = require("jsonwebtoken");
+const secret = process.env.JWT;
 module.exports = router;
+
+const authRequired = async (req, res, next) => {
+  // We grab the token from the cookies
+  const token = req.headers.authorization;
+  // jwt verify throws an exception when the token isn't valid
+  try {
+    const { id, adminStatus } = await jwt.verify(token, secret);
+    req.userId = id;
+    req.adminStatus = adminStatus;
+  } catch (error) {
+    res.status(401).send({
+      loggedIn: false,
+      message: "Unauthorized",
+    });
+    return;
+  }
+  next();
+};
 
 // GET /api/products (gets all products)
 router.get("/", async (req, res, next) => {
@@ -47,52 +67,67 @@ router.get("/:productName", async (req, res, next) => {
 
 //POST /api/products (creates a new product)
 
-router.post("/", async (req, res, next) => {
+router.post("/", authRequired, async (req, res, next) => {
   try {
-    let newProduct = await Product.create(req.body);
-    res.status(201).json(newProduct);
+    // only admins can create new products
+    if (req.adminStatus) {
+      let newProduct = await Product.create(req.body);
+      res.status(201).json(newProduct);
+    }
   } catch (err) {
     next(err);
   }
 });
 
 // DELETE /api/products/:id (delete product)
-router.delete("/:id", async (req, res, next) => {
+router.delete("/:id", authRequired, async (req, res, next) => {
   try {
-    await Product.destroy({ where: { id: req.params.id } });
-    res.sendStatus(204);
+    // only admins can delete products
+    if (req.adminStatus) {
+      await Product.destroy({ where: { id: req.params.id } });
+      res.sendStatus(204);
+    }
   } catch (err) {
     next(err);
   }
 });
 
 // PUT /api/products/:id  (update product by id)
-router.put("/:id", async (req, res, next) => {
+router.put("/:id", authRequired, async (req, res, next) => {
   try {
-    let updatedProductInfo = await Product.findByPk(req.params.id);
-    res.json(await updatedProductInfo.update(req.body));
+    // only admins can edit products
+    if (req.adminStatus) {
+      let updatedProductInfo = await Product.findByPk(req.params.id);
+      res.json(await updatedProductInfo.update(req.body));
+    }
   } catch (err) {
     next(err);
   }
 });
 
 // PUT /api/products/:id  (inc product amount by id)
-router.put("/:id/increase", async (req, res, next) => {
+router.put("/:id/increase", authRequired, async (req, res, next) => {
   try {
-    let product = Product.findByPk(req.params.id);
-    product.stockQuantity++;
-    res.json(await product.save());
+    // only admins can edit products
+    if (req.adminStatus) {
+      let product = Product.findByPk(req.params.id);
+      product.stockQuantity++;
+      res.json(await product.save());
+    }
   } catch (err) {
     next(err);
   }
 });
 
 // PUT /api/products/:id  (dec product amount by id)
-router.put("/:id/decrease", async (req, res, next) => {
+router.put("/:id/decrease", authRequired, async (req, res, next) => {
   try {
-    let product = Product.findByPk(req.params.id);
-    product.stockQuantity--;
-    res.json(await product.save());
+    // only admins can edit products
+    if (req.adminStatus) {
+      let product = Product.findByPk(req.params.id);
+      product.stockQuantity--;
+      res.json(await product.save());
+    }
   } catch (err) {
     next(err);
   }
